moved more code to test words
